+++ src/ExpoYandexMapKitView.web.tsx
@@ -20,6 +20,7 @@ export const YandexMapView = React.forwardRef<YandexMapViewRef, YandexMapViewPro
         getVisibleRegion: unsupported,
         getScreenPoints: unsupported,
         getWorldPoints: unsupported,
+        fitToCoordinates: unsupported,
       };
     }, []);
     return null;